codeblocks cutoff - fixes #1
Last line of code blocks would sometimes be cutoff.

Issue is that sometimes on a new start of Chrome/Chromium, each class=code
block would report eactly the number of lines in its text area, but every other
time it would also include a blank line at the end.

So, the previous fix was to always remove the last line.  However, it would
sometimes (rarely!) cause the text to be cutoff if that extra blank line wasn't
there.

So, we now only cut the last line if its blank.