make tidy - fixed tab checking, and made better!
Old approach to checking tabs on code blocks was to assume all code is
tab-indented.  But, duh, this doesn't work, as tidy won't output tabs (converts
all to spaces).

New way is to keep all indenting as spaces.  The new better check is that we
pipe all code blocks through the python syntax checker, to get a warm fuzzy
that things are pretty good.

Note - this requires that the javascript has to convert code from spaces to
tabs in the browser (not yet implemented).